change relative addressing mode
Because of the ZeroPageRelative addressing mode that needed to perform 2
actions on the memory, the Relative addressing mode got its resolution
privilege revoked. Now, the Relative addressing mode does return None as
resolution and the resolution has to be done by hand in the according
instructions.
The ZeroPageRelative addressing mode resolves into the zero page
address.